+++ CHANGELOG.md
@@ -1,1 +1,13 @@
-
+# v0.2.0
+## Added filters
+- Added Twig style filters in the spirit of PHPlater
+- Multiple filters can be applied
+- The filter separator can be defined by user
+- If template variable is array, json or object, filters can be applied to it as well
+# v0.1.0
+## Initial release candidate
+- Basic templating functionality
+- Nesting of templates
+- User defined delimiters and tags
+- Template can be file or string
+- Varables added can be arrays, json and objects, and data can be retrieved by chaining